Allow running docker image as non-root user
The docker image expected to be root as it was doing operations
that require root permissions in the entrypoint.sh (so at runtime).

Some users might want to run the entire container as a non-root user,
so let's cater for both scenarios by adding a USER instruction in the Dockerfile
and adding the group/user there as well.

By default this will mean we now run the container as a non-root user, but if
a user still wants to run as root that is possible by specifying `--user root`
using docker cli or whichever equivalent in other container runtimes.

This also means we don't need the gosu command, so removed it from entrypoint.sh
and the Dockerfile.